+++ plugin-dev/Source/Sentry/Private/Mac/MacSentrySubsystem.cpp
@@ -9,6 +9,7 @@
 #include "SentrySettings.h"
 
 #include "Utils/SentryFileUtils.h"
+#include "Utils/SentryScreenshotUtils.h"
 
 #include "Misc/CoreDelegates.h"
 #include "Misc/FileHelper.h"
@@ -42,41 +43,11 @@ TSharedPtr<ISentryId> FMacSentrySubsystem::CaptureEnsure(const FString& type, co
 
 FString FMacSentrySubsystem::TryCaptureScreenshot() const
 {
-	NSWindow* MainWindow = [NSApp mainWindow];
-	if (!MainWindow)
+	const FString ScreenshotPath = GetScreenshotPath();
+	if (!SentryScreenshotUtils::CaptureScreenshot(ScreenshotPath))
 	{
-		UE_LOG(LogSentrySdk, Error, TEXT("No main window found!"));
 		return FString("");
 	}
-
-	NSRect WindowRect = [MainWindow frame];
-	CGWindowID WindowID = (CGWindowID)[MainWindow windowNumber];
-	CGImageRef ScreenshotRef = CGWindowListCreateImage(WindowRect, kCGWindowListOptionIncludingWindow, WindowID, kCGWindowImageDefault);
-
-	if (!ScreenshotRef)
-	{
-		UE_LOG(LogSentrySdk, Error, TEXT("Failed to capture screenshot - invalid ScreenshotRef."));
-		return FString("");
-	}
-
-	NSBitmapImageRep* BitmapRep = [[NSBitmapImageRep alloc] initWithCGImage:ScreenshotRef];
-	NSData* ImageData = [BitmapRep representationUsingType:NSBitmapImageFileTypePNG properties:@{}];
-
-	TArray<uint8> ImageBytes;
-	uint32 SavedSize = (uint32)[ImageData length];
-	ImageBytes.AddUninitialized(SavedSize);
-	FPlatformMemory::Memcpy(ImageBytes.GetData(), [ImageData bytes], SavedSize);
-
-	CGImageRelease(ScreenshotRef);
-
-	FString ScreenshotPath = GetScreenshotPath();
-
-	if (!FFileHelper::SaveArrayToFile(ImageBytes, *ScreenshotPath))
-	{
-		UE_LOG(LogSentrySdk, Error, TEXT("Failed to save screenshot to: %s"), *ScreenshotPath);
-		return FString("");
-	}
-
 	return ScreenshotPath;
 }